failing test demonstrating padding oracle attack
We control the last byte of the decrypted plaintext by XOR'ing
the last byte of the IV with the correct padding byte, followed
by the desired plaintext byte. Setting the last byte of the
plaintext to 1 causes this to be interpreted as a valid one-byte
padding, resulting in the signature check failing. Setting it to
a value other than 1 or the previous value causes the padding to
be invalid and for OpenSSL to raise an error.